screens refactor to views
