--- lib/screens/cart_screen.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CartScreen extends StatefulWidget {
-  const CartScreen({Key? key}) : super(key: key);
-
-  @override
-  State<CartScreen> createState() => _CartScreenState();
-}
-
-class _CartScreenState extends State<CartScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold();
-  }
-}
