--- lib/screens/search_screen.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SearchScreen extends StatefulWidget {
-  const SearchScreen({Key? key}) : super(key: key);
-
-  @override
-  State<SearchScreen> createState() => _SearchScreenState();
-}
-
-class _SearchScreenState extends State<SearchScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Center(
-        child: Text("Search History"),
-      ),
-    );
-  }
-}
