+++ lib/views/pharmacy/pharma_view.dart
