+++ lib/views/splash_screen.dart
